Backend: Fix bug of API `consumer/comment/xxx-media`

+++ backend/sustech-regency-db/src/main/java/com/sustech/regency/db/po/Exhibitable.java
@@ -11,6 +11,6 @@ public interface Exhibitable<Display> {
 	void setMediaId(String mediaId);
 	String getMediaId();
 
-	void setDisplayId(Integer displayId);
-	Integer getDisplayId();
+	void setDisplayId(Long displayId);
+	Long getDisplayId();
 }
